test(nomad): write test data and verify on another node
In this test a job is created which writes a test file of 1M to a Hetzner Cloud volume. Afterwards, this job is deleted and another job on another node is started, that verifies that the file is present and has a size greater than zero.

+++ test/e2e/nomad/utils.go
@@ -3,6 +3,7 @@ package e2e
 import (
 	"context"
 	"fmt"
+	"math"
 	"os"
 	"sync"
 	"time"
@@ -11,6 +12,8 @@ import (
 
 	"github.com/hetznercloud/hcloud-go/v2/hcloud"
 )
+
+const InitialVolumeCapacity = 10737418240 // 10GiB
 
 type Cluster struct {
 	hcloudClient *hcloud.Client
@@ -82,6 +85,11 @@ func (cluster *Cluster) Cleanup() []error {
 		vol, _, err := cluster.hcloudClient.Volume.GetByName(context.Background(), volName)
 		if err != nil {
 			cleanupErrors = append(cleanupErrors, err)
+			continue
+		}
+		if vol == nil {
+			cleanupErrors = append(cleanupErrors, fmt.Errorf("volume %s not found on hcloud", volName))
+			continue
 		}
 		_, err = cluster.hcloudClient.Volume.Delete(context.Background(), vol)
 		if err != nil {
@@ -118,4 +126,106 @@ func (cluster *Cluster) DeleteVolume(externalVolID string, w *nomad.WriteOptions
 	delete(cluster.volumesCreated, externalVolID)
 
 	return nil
+}
+
+func (cluster *Cluster) ExecInAlloc(alloc *nomad.Allocation, task string, command []string) (int, error) {
+	exitCode, err := cluster.nomadClient.Allocations().Exec(
+		context.Background(),
+		alloc,
+		task,
+		true,
+		command,
+		os.Stdin,
+		os.Stdout,
+		os.Stderr,
+		make(<-chan nomad.TerminalSize),
+		&nomad.QueryOptions{},
+	)
+	if err != nil {
+		return exitCode, err
+	}
+	return exitCode, nil
+}
+
+func (cluster *Cluster) WaitForRunningJob(jobID string) (*nomad.AllocationListStub, error) {
+	for retry := range 10 {
+		allocs, _, err := cluster.nomadClient.Jobs().Allocations(
+			jobID,
+			false,
+			&nomad.QueryOptions{},
+		)
+		if err != nil {
+			return nil, err
+		}
+
+		for _, alloc := range allocs {
+			if alloc.ClientStatus == "running" {
+				return alloc, nil
+			}
+		}
+
+		backoffSleep(retry)
+	}
+	return nil, fmt.Errorf("no running allocation for job %s", jobID)
+}
+
+func createVolumeSpec(id string) *nomad.CSIVolume {
+	return &nomad.CSIVolume{
+		ID:                   id,
+		Name:                 id,
+		Namespace:            "default",
+		PluginID:             "csi.hetzner.cloud",
+		RequestedCapacityMin: InitialVolumeCapacity,
+		RequestedCapabilities: []*nomad.CSIVolumeCapability{
+			{
+				AccessMode:     "single-node-writer",
+				AttachmentMode: "file-system",
+			},
+		},
+		MountOptions: &nomad.CSIMountOptions{
+			FSType: "ext4",
+			MountFlags: []string{
+				"discard",
+				"defaults",
+			},
+		},
+	}
+
+}
+
+func createBusyboxWithVolumeJobSpec(id string, volumeID string, mountPath string) *nomad.Job {
+	job := nomad.NewServiceJob(id, id, "global", 50)
+	taskGroup := nomad.NewTaskGroup(id, 1)
+
+	taskGroup.Volumes = map[string]*nomad.VolumeRequest{
+		volumeID: {
+			Name:           volumeID,
+			ReadOnly:       false,
+			Type:           "csi",
+			Source:         volumeID,
+			AttachmentMode: "file-system",
+			AccessMode:     "single-node-writer",
+			PerAlloc:       false,
+		},
+	}
+
+	task := nomad.NewTask(id, "docker")
+	task = task.SetConfig("image", "busybox:stable")
+	task = task.SetConfig("command", "sleep")
+	task = task.SetConfig("args", []string{"3600"})
+
+	task.VolumeMounts = append(task.VolumeMounts, &nomad.VolumeMount{
+		Volume:      &volumeID,
+		Destination: &mountPath,
+	})
+
+	taskGroup = taskGroup.AddTask(task)
+	job = job.AddTaskGroup(taskGroup)
+	return job
+}
+
+func backoffSleep(retry int) {
+	delay := math.Pow(2, float64(retry))
+	delay = math.Min(delay, 16)
+	time.Sleep(time.Second * time.Duration(delay))
 }
